fix: Correctly parse host with port
This patch improves the image name parsing to account for hosts with
ports as well as multiple '/' in the image name.

Note that while a similar end result could be achieved with a regex,
the resulting code would likely be too terse for anyone without a
reasonable grasp of the docker ecosystem (tag formats etc) to understand
quickly. Lets stick to verbose and readable, as it is not a performance
concern in this case.

Co-authored-by: Charlie Briggs <charlie.briggs@ft.com>