fix: fix 'inspect an image with unspported pkg manager' test
'inspect an image with an unsupported pkg manager' is currently
broken on master. Here is a PR that merely adds some text
to README.md that also fails on CI due to the aforementioned test.

https://github.com/snyk/snyk-docker-plugin/pull/62

The test fails due to incompatible error strings, nothing major.

I've applied a poor fix to it for now, just to fix CI. We should
investigate this and properly fix it.