fix: Bug with flaky findGlobs results
In app scanning we had an onerous issue where we would sometimes return
empty results for manifest files, even though files were present on
image. This was narrowed down to the issue of error handling inside
runSafe().

Therefore, update runSafe() not to nullify stdout/stderr result, even if
some command resulted in error code. This shouldn't change existing
clients' behaviour, except for the case of listing multiple files ("ls
-R /"). In that case some files might become inaccessible during long
operation and make the command return error code. In this case, we
return stdout/stderr to the caller, so the result can be parsed
normally.