chore: refactor the apk analyzer to decouple docker from the logic
decouple the way the retrieve the apk data (apk DB file) which is done by running "docker run ..." from the logic that acts on the file's content.
this will allow reusing the logic to analyse a file acquired in a different way (static scanning) while still supporting the existing way.

right now this does absolutely nothing except for showing an initial proposal of how the rest of the analysers might look like.