fix: decouple image analyzers and introduce static scanning opts
Decouple the way we retrieve package manager files (e.g. APK DB file), which is currently done with Docker run, from the logic that processes the file content. This will allow reusing the logic of processing a file that was acquired in different ways (e.g. statically or dynamically with Docker) while still supporting the existing functionality.

Additionally introduced static stanning options which serve as hints that the caller wants a static analysis. The code now branches at the entrypoint "inspect()" for the two new paths: static analysis and dynamic (Docker) analysis.