fix: resolve paths in extracted files in docker-archive
The expectation until now was that files are extracted as "some/path/to/file.xyz".
Notice it didn't start with "/" so we prepended that to the filename.

However, using a RHEL image, files are named "./some/path/to/file.xyz", and if we just prepend "/" the path becomes "/./some/..." which will never match.

This fix now properly resolves paths by using path.resolve("/", <the path>") so that we get consistent paths regardless of relative paths.