fix: allow rpm data extraction to fail on BusyBox
today we allowm rpm data extraction (package list) to fail only if rpm is not installed in the image.
on BusyBox images, rpm is installed, but on a very minimal version that does not support its usage.
furthermore, while it may be used to install packages, it probably can't list packages.

this commit extends our tolerance to rpm failures to include BusyBox's failure, manifesting in an error message about wrong flags and having some mentions of BusyBox and the multi-call binary.