fix: organise and document the binaries portion of analysis parsing
parseAnalysisResults() is used in both dynamic and static scanning flows.
in the dynamic scanning flow, we expect the analysis to have a ".binaries" property, which we then return.
in the static scanning flow, we implicitly return "undefined" in its place because the analysis won't have the binaries property.

this commit surfaces these gaps and improves the types around it.