feat: return a new, empty scanResults array
the `scanResults` array contains unstructured data for different ecosystems encountered while scanning an image.
the `data` is expected to be consistent for each `type`+`version` couple.

the existing `package` object will now begin deprecation, and should be swallowed in the new `scanResults` object as one of the `package`es that we found when scanning an image.
(that is; scanning an image may result in a package for OS data, two packages for Node data, a package for Java, etc.)