feat: allow complex file matching when scanning files statically
Introduce a "match" function that takes a path and returns a yes/no result whether the path matches what we are looking for in the specific extracting step/action. This allows us to not only "match" on paths but also to perform more complex actions like filtering/excluding things we don't care about.