fix: allow files to be extracted by directly matching on their name
Remove the old use of matching by a pattern and fix a bug where we were looking at the list of extracted files and trying to match by the pattern directly. In patterns like "**/node" it would never work as we would be comparing an absolute path (e.g. /var/lib/node) to the pattern ("**/node") as a direct match.