feat: return multiple scan results to support application dependencies
BREAKING CHANGE: this commit changes the return type of the main `inspect` function to adhere to Snyk's CLI's newer format that supports multiple scan results, sometimes called projects.
To use the new format, iterate over the new `scannedProjects` property of the return value. Each item has a `depTree` with dependencies (similar to the old `Package`) and a `packageManager`.