test: add tests to assert transitive deps are not found with RPM
Shows that there is a bug/limitation in the way we produce the dependencies for RPM projects.
Currently the dependencies are a flat list rather than a dependency tree or graph.

This means that transitive dependencies introduced by installing a package are not linked back to the Dockerfile instruction that introduced them.