feat: add flag imageNameAndTag to preserve this data in archive scans
Override the "name" and "version" fields of the OS dependencies result when scanning image archives.

This is NOT used by the Snyk CLI!
It is used by K8s-Monitor and DRA to preserve the image identifier when scanning archives.
The archives do not contain any data about the image's name and tag (since they are only known by the container registry) and in some contexts we know this data and want to keep it.