chore: fix flaky static scanning test with debian:10
Debian:10 was updated a few dayss ago. It seems that there was some sort of dependency on an earlier test where a distroless image was pulled that also depended on debian:10. Since the image is now different, the flaky test assumed it was present locally and called directly "docker save". This failed because the images between the two tests are now different. One solution is to use "docker pull" before calling "docker save".

However, static analysis is now able to analyze images directly from the container registry, so instead the test is modified to call the plugin directly with the image name and tag.