feat: manifest files contents are now returned as a ScanResult Fact
Having the manifest files returned at the top-level plugin response does not conform with other Snyk CLI plugins.
The manifest files have only two use-cases: kubernetes-monitor and DRA.

We prefer to "hide" the manifest files as a Fact in the OS dependencies ScanResult (the 1st ScanResult).
These manifest files need further processing so we can extract actual package manager information and to create projects in Snyk.

Additionally, as part of this change, change the interface of the manifest files to return their contents as a Base64 string.
This is needed for two reasons:
- the existing Buffer interface cannot cleanly be transferred over the network (considering we use JSON in our APIs)
- we cannot make assumptions about the file encoding otherwise we can accidentally corrupt the data we receive