feat: say which files were tested when producing npm/yarn analysis
This is a feature already supported in the backend and is used when importing Node.js Projects. It is generated by the backend as part of a manifest file analysis.

This feature we previously got for free since Node.js scans were done on a manifest file. However, now that we produce dependencies directly we need to bring this piece of information as part of the analysis to preserve backwards compatibility.