fix: annotate packages with the instruction that introduced them
This fixes a regression with rolling out static scanning. Previously dependencies used to be annotated with the Dockerfile instruction that introduced them. Without this fix we could not properly distinguish between packages installed by the base OS and packages installed by user instructions in the Dockerfile. We were incorrectly representing all packages as coming from the base image.