fix: allow base image to be missing if it cannot be read correctly
In some cases where the base image in the Dockerfile's FROM instruction contains arguments that are substituted at build-time, we would incorrectly detect the base image as "null:null" which is a bad user experience. We fix this by ensuring that we report the base image only if we can read every variable in the FROM instruction.