refactor: Update expectations for images
The image `debian:10` was modified on
the public repo, causing a test to fail.
Our test expectation has been updated to match
the new correct value, and pinned to a specific
hash.

Similarly, the image `opensuse-leap:15.1` was modified on
the public repo. This change affected the number
of dependencies, causing a test to fail.
So our test expectation has been updated to match
the new correct value.