fix: ensure to capture and rethrow errors from image layer parsing
Right now any errors that occur during image layer parsing are not properly caught (because they occur either in an async or EventEmitter context) and result in uncaught exceptions.

One of these places is the Go binaries parser, which we now properly try-catch and reject the promise that wraps it.
Additionally we need to try-catch the image layer parsing process as a whole, which was missing before.