fix: best-effort attempt to read OS release files
This is instead of the current implementation of failing on the first OS release parser.
Some images may contain multiple OS release files. One of them would be malformed, whereas the other would contain the data for the OS release.

This change ensures all handlers get the chance to run and try and parse the OS release.