fix: Update snyk-docker-pull to fix exception
Updated snyk-docker-pull which has an updated version of
docker-registry-v2-client that fixed an issue when getting multiarch
 manifest by its digest returning a manifest list, which resulted in a
'cannot read property of undefined' exception, when trying to access the
manifest.config.digest object (as there is no config object in manifest
list).
The latest fix in the v2-client lib makes sure the getManifest always
returns a manifest.v2 object rather than a list (in case a list is returned
by the API call the default linux/amd64 arch is chosen and returned by it).