fix: allow base image to be partially missing
In some cases where the base image in the Dockerfile's FROM instruction contains arguments that are substituted at build-time, we would incorrectly detect the base image as /image:tag which is a bad user experience. We fix this by ensuring that we report the base image only if we can read every variable in the FROM instruction.