feat: Find and unpack shaded jars
If nested jars found, we are not returning the parent (fat / shaded) jar. We are hiding unpacking jar functionality behind CLI flag. The uberjar code wont run at all if app-vulns is not specified. In order to make this clear to the user, we're throwing an error in case
the user specifies --shaded-jars, but does not specify --app-vulns.