feat: Add image creation time fact
As part of the rule-based auto-import from container registries,
we would like to base some of the import rules on the image creation
time, found in the image config. Since the image creation time is
not a required field in the spec, we would like to examine what is the
percentage of images that have this information missing. Adding this
attribute as a fact will allow us to collect this information.