test: refactor tests to use describe.each
we decided that shaded-jars-depth isn't the correct terminoligy, and we
want to deprecate it. first step was adding nested-jars-depth. so currently
we have duplicate tests to make sure both flags are functioning.
instead of duplicating every test we add or modify, I've refactored
to use describe.each, so we can amend/add any test only once.