fix: ensure large objects are not loaded in memory when parsing them
Whenever we attempted to parse oci-archive files, we first loaded any file we encountered in memory fully before we JSON-parsed it.
Some files may be really large and may consume lots of memory.

Introduce a hard limit to read only up to 2 MiB of the file and to attempt to parse the result.