fix: Fixed invalid jar fingerprints for maven projects
Refactored the recursion of unpacking nested JARs to avoid mutating objects
which was difficult to follow and caused incorrect dependency allocation
in the fingerprints which resulted in an invalid dep-graph.

We are now including JARs that have nested JARs as we now look for
dependencies via pom.properties in each JAR; omitting would mean we
could be missing vulns.