fix: private jars with no root artifact
In some cases, we can't find any artifact ID in any po.properties file
inside of a jar file, that matches the jar path. In this case we
couldn't create a dep graph since it had no root. This fix adds some
default data so that we can create a dep graph and show dependencies and
find vulns.