feat: remove docker experimental check [MYC-117]
This commit removes the check if Docker is running with experimental
features enabled, which was required previously to do platform-specific
image pulls through the `--platform` flag to `docker pull`.

If the Docker binary should not have support for the platform flag at
all, we return an error message that it needs to be upgraded to at least
17.12. From 17.12 to 20.10, the platform flag was considered
experimental, and if the user should not have the experimental features
enabled, we return the error message from Docker that it needs to be
enabled.

For versions above 20.10 or above 17.12 with experimental features
enabled, we simply run `docker pull --platform=...` if the platform was
defined.

It also fixes a small issue with how we passed the arguments to the
`docker pull` command, which if the `platform` flag was not needed, we
passed an empty string to `subProcess.execute`. This worked because
`subProcess.execute` seems to skip empty arguments, else it would have
called `docker pull "" <image>`, which would fail. But as this is an
implementation detail that is not guaranteed to remain, this commit
improves this by not passing an empty string at all.