fix: streamToString() correctly building and encoding the string
There is a bug where the string is built as a collection of chunks, but the chunks are encoded as they are concatenated.
When using any multi-byte encoding such as Base64 or UTF-8 this means that we may receive a chunk of bytes with incomplete characters. By encoding it too early we get the wrong string.

The fix is to now encode the whole string after all the chunks are collected.