test: fix regression test shell issues
`==` is a bashism, `=` is the shell equivalent. Interestingly, when run
on MacOS, `==` works as well, which is why local testing didn't bring up
this issue.

Further, `@*` concatenates the array of arguments into a single variable
seperated by the `IFS`, which is a space. So the end result should be
the same.

These changes have been tested in the same Docker image that is used in
CI.