feat: support trimmed, stripped and CGo binaries
This commit updates our support for Go binaries built with the
`-trimpath` flag, stripped binaries (`-ldflags="-s -w"`), vendored
dependencies or with CGo enabled. This means that customers can finally
scan pretty much any Go binary (built with the Go compiler) and get
meaningful results 🎉

The new support is implemented by parsing the PCLN table and going
through the list of files that is in there. This list contains the paths
of the files which includes their module and package name. We extract
this to match them to the Go module version that we get from the
`.go.buildinfo` ELF section (same as what `go version -m <binary>`
does).

The actual implementation needs to check for a couple of edge-cases, as
trimmed binaries or vendored dependencies need to be treated slighly
differently than regular binaries.

The tests have been extended to support that as well, a vendored
binary has been added to the tests and the regression tests will always
build vendored binaries as well.