fix: handle Go binaries from the Go distributions
Binaries that are built in the Go distribution (e.g. the Go compiler,
`gofmt`, `vet` and many more) do not have a `mod` directive like
"normal" Go binaries. This seems to be a special case for all binaries
from the Go source / standard library. So far, our parsing logic
couldn't handle the missing `mod` directive, which means that all these
binaries currently have the same project name `-compiler=gc`.

This commit fixes that behaviour by naming the binaries with a fixed
prefix `go-stdlib@` plus the path that we can extract from the `path`
directive in the build metadata. This results in names like
`go-stdlib@cmd/vet`.

Note that this is only relevant for the CLI output, a `snyk container
monitor` does not seem to use the project name from the CLI, but rather
the path.