fix: skip CGo binaries without PCLNTable
Some CGo-built binaries might not contain a pcln table, as is described
in [Building a better Go linker](https://docs.google.com/document/d/1D13QhciikbdLtaI67U6Ble5d_1nsI4befEd6_k1z91U/view):

> For “cgo” binaries, which may make arbitrary use of C libraries, the
Go linker links all of the Go code into a single native object file and
then invokes the system linker to produce the final binary.

One example and the reason why we have found this issue is containerd,
that is built with CGo enabled.

I do not know how exactly to build a Go binary without a PCLN table, but
somehow it must be possible :-)

So far, we have thrown an error and thus exited the scan if the PCLN
table was missing in a Go binary, because we did not expect this case to
exist. Now that we know it exists, this commit fixes that behaviour and
simply skips the binary in the scan.