fix: amending condition for 'manifest unknown' errors
This commit amends the condition based on the latest response from Dockerhub when the manifest for a given image cannot be found. It ensures that the errors is not treated as a generic error and will not fall back to use snyk-docker-pull and fail with invalid credentials error when using the docker binary.