fix: No HTTP_PROXY when connecting to Docker socket
CLIv2 sets up a global proxy where all traffic from the typescript
modules / CLI parts are routed through a Go proxy. This is an issue for
the Docker modem where we want to connect to Docker's socket directly,
as the proxy alters the behaviour of the calls. Because traffic to a
socket does not need a `host` or `port` to be set, the Go proxy that got
that traffic re-routed it to `localhost:80`, which results in the CLI
failing.

I've opened an [upstream PR in the `global-agent`](https://github.com/gajus/global-agent/issues/59)
repo to not proxy traffic that should reach sockets in the first place.
However, until this is fixed, we should simply use a custom agent for
docker modem.