fix: go binary file path determination
A Go binary does not contain a list of packages in it's metadata, but we
need it to query our VulnDB correctly. We're instead extracting this
information from the list of files in the PCLN table. The algorithm to
extract the list of packages from these files has proven to be buggy, we
got a report of a user attempting to scan `kyverno` and not getting the
correct dependencies.

This bug was triggered by the following three things in combination:
- The binary was built with `trimpath`
- The binary depended on `golang.org/x/net`
- The standard library also depends on a vendored copy of
  `golang.org/x/net`

Now the file list contained entries that looked something like this:
```
golang.org/x/net/http/http.go
vendor/golang.org/x/net/http/http.go
github.com/spf13/cobra/cobra.go
```

The issue was that we tried to find the correct module cache path for
every file. When we hit the `vendor/...` file, the module cache path was
assumed to be `vendor/`, which means that all following files were
excluded.

This issue does not show up if the binary does not depend on any of
these vendored modules from the standard library, as they did not match
any module and where thus skipped.

To fix this issue, this commit fixes the way we determine these paths.
We now get these paths before looping over the files, and also determine
whether a binary is trimmed or not more accurately.

As a bonus, this PR means that we can also handle binaries that have
mixed vendored and non-vendored dependencies. This doesn't exist
currently, but there are proposals out there to extend the Go build
command to support it.