refactor: rename depType / packageManager to packageFormat
This commit renames the depType fields and variables, as well as some
weird occurrences of `packageManager` to be what it really is, the
`packageFormat`. For most package managers, this would be the same.
However, `apt` uses `.deb` as their package format, so this commit
should make it easier to distinguish between the two.