feat: redhat module & repository support
This commit makes use of the added PURL-support in DepGraph to include
the RPM package modules as the "namespace", and the repositories as part
of the qualifiers in the PURL.

With this, Vuln can filter down vulnerabilities based on the
repository-to-cpe mapping and also query packages by module. Support for
this has been added through [Vuln #708](https://github.com/snyk/vuln/pull/708).