fix: unify error message formatting
This commit removes all occurrences of printed errors that are being
formatted with `JSON.stringify()`, as that formats the error to an empty
object `{}`, which is not helpful.

It also removes calls to `debug` with `error` as the second argument,
and embeds them into the string. This is purely to make all our error
messages look the same.