chore: reproduce windows scanning issues
Customer has reported a bug with scanning some Go binaries on windows.
Interestingly, not all binaries seem to fail, only some. For the
failing binaries, no dependencies are being found.

This commit adds a test that runs on Windows to make sure that the
scanning works. The snapshot has been generated by running the test on
MacOS, which illustrates the point that they _should_ produce the exact
same snapshot as the Windows one.

The only thing that needed changing was the filepaths, e.g. from
`"targetFile": "/livenessprobe"` to `"targetFile": "\\livenessprobe"`.