fix: use path.posix for Go file / module lookup
This commit switches the usage of the `path` module to use `path.posix`
instead, fixing an issue where when the CLI was running on Windows, the
Go dependency resolution returned no results.

The underlying issue is that the Go binary metadata always contains
normal slashes in it's path, while our usage of `path.join` and
`path.parse` meant that we were constructing paths with backward-slashes
instead.