fix: error handling when pulling images with pull library
So far, we've only handled errors when pulling with the Docker CLI. With
the Docker CLI, we've "translated" errors to make them more helpful.

The Windows tests have now started to fail because of a broken Docker
installation, as we then fell back to the `snyk-docker-pull` library,
which didn't have that error translation.

This commit fixes that, and applies that error handling to both the CLI
errors as well as errors from snyk-docker-pull.