fix: removes meta-common-packages generation for Linux package managers
Frequently found dependencies can cause many upgrade paths to be generated which creates a bad
UX when issues cards are displayed in the UI. Grouping these packages under a fake package,
meta-common-packages solved this but broke the upgradable logic for Linux package managers.

This commit attaches frequently fond deps directly to the tree route for Linux package
managers so the upgrade path shows the package as a direct dependency and is upgradable.