fix: ignore composer lock files that are invalid
Sometimes (like in drupal 10), empty "composer.lock" files are created. This change skips over files that are invalid, such as when the "packages" property is missing.

The original issue we encountered is an empty json object in the lockfile and manifest. The error message specifically said it was looking for a "packages" property in the object.
We considered having the exception comparison to be more specific to lessen the number of potential false negatives. However, that would require either checking the error message which is unreliable (due to coming from another library), or checking the contents of the file, which is not the job of snyk-docker-plugin, but of composer-lockfile-parser.