fix: add support for deleted folders
When a file or folder is deleted, an entry that starts with .wh is created.
Up until now, Snyk detected files that were deleted, but not folders.

This change adds support for deleted folders by checking whether a file path starts with the path of the deleted folder. Meaning if we know that "/app/dependencies" was deleted, then a file path like "/app/dependencies/jar/library.jar" will be marked as deleted as well.