fix: uses options.platform to select correct manifest from OCI layout
When selecting the correct manifest from an OCI layout, use the platform
supplied in options to match against an index or fallback to only
manifest in the case of single manifest existing in layout. This ensures
multi-platform images saved with containerd get teh correct manifest and
sets the platform correctly from the associated config.