chore: declare node version
Align this repo with the CLI node version.

The .node-version file is used in Node.js projects to specify the
version of Node.js required for the project. It helps developers
and automated tools like version managers (e.g., nvm, fnm) to ensure the
correct Node.js version is used.

More information
https://github.com/shadowspawn/node-version-usage