fix: handle removal of multiple additions UNIFY-243
The issue was that in the extracted layers we could find multiple additions of the exact same file.
Docker allows the addition of exactly the same file in exaclty the same path multiple times.
We parse layers in reverse chronological order, so encountering removals first means those files are 100% removed.
For clarifying, removals are marked with ".wh.".
Before this change, we would only take into account a removal for the first instance of that exact same file in the following layers.
In this comit I have also added 2 tests and 2 fixtures for those tests.