fix: optional python dependencies
Optional Python dependencies are being over connected.

When building Python dep-graphs the current algorithm does not consider
whether a transitive should be traversed based on 'extra' definitions.

When a Python dependency uses extras, this means that dependency is
optional and will only be installed when the installer asks.

For example if 'my-package' has the following METADATA snippet:

```
Requires-Dist: logger>=1.0
Provides-Extra: test
Requires-Dist: tester>=0.5.0; extra == 'test'
```

This means 'logger' is always installed, but 'tester' is only installed
if 'my-package' is installed with extra 'test' dependencies, like so:

```
my-package[test]==1.2.3
```

The current algorithm sometimes gets away with this lack of accuracy
because the dependency is skipped if the METADATA file doesn't even
exist. However in many cases what's optional in one transitive line is
not in another and this can mean the METADATA file does exists. This
results in the current algorithm accidentally associating potentially
large sub-graphs to transitive lines that should be terminated when
extras are not being used.

The first change here introduces code that firstly parses out these
extra definitions in both requirements.txt and METADATA files. Then also
parsers that pick out the Provides-Extra and Requires-Dist extra
environment markers.

The second change adapts the pip dep-graph builder to take into account
extra properties to decide whether a dependency should be traversed or
not.