fix: update quay tests and add a script to update them (#708)
StreamOS 9 and StreamOS 10 tests constantly fail in snyk-docker-plugin.

This is because Redhat’s quay.io repo does not store old shas once an image is updated. So every time they update the image, the hardcoded sha values need to be updated

To alleviate this, I’ve written a shell script to update the shas and update the test snapshot

So if these fail in the future, run:
npm run update-quay-tests