fix: cap stream size to node.js max buffer size (#715)
Now that we can handle scanning larger files, there's a very small chance of a failure if the file size is greater than the max file size allowed by node.js AND it has an ELF header

The issue here is it can fail for large ELF files (> 4GB). The chance if this is near zero, I'm not aware of any go modules this large, but if someone has another executable that is this large (still pretty unlikely), it could crash the scan. So I've just added one more check